+++ old_project/APP_GPS_config.py
@@ -1,9 +1,9 @@
-import pm_g5t as pm_sensor
-import th_htu21d as tmp_sensor
-import light_bh1750fvi as light_sensor
-import gas_tvoc_sgp30 as gas_sensor
+import lib.pm_g5t as pm_sensor
+import lib.th_htu21d as tmp_sensor
+import lib.light_bh1750fvi as light_sensor
+import lib.gas_tvoc_sgp30 as gas_sensor
 import pyupm_i2clcd as upmLCD
-import gps_neo6m as gps
+import lib.gps_neo6m as gps
 
 Version = "5.2a.2"
